refactor: Move viewshed function to separate files
